+++ pkg/network/cache/clients.go
@@ -372,6 +372,7 @@ func (x *connections) ObjectRangeInit(ctx context.Context, cnr cid.ID, id oid.ID
 	var res *client.ObjectRangeReader
 	return res, x.forAny(ctx, func(ctx context.Context, c *client.Client) error {
 		var err error
+		//nolint:staticcheck
 		res, err = c.ObjectRangeInit(ctx, cnr, id, off, ln, signer, opts)
 		return err
 	})